Add [AsyncAlert] component
This commit adds [AsyncAlert] component, which is just a Promise wrapper around
react-native's [Alert] component. This way, we can be in control of the
state depending on the user input. This is needed, for example, when trying to
prevent navigation to a new screen if user cancels the action, or, if
different arguments need to be passed depending on the input.